set home cursor theme to DMZ White
by default a nix application will choose some default cursor
theme; this becomes problematic on large screens with small
pixels. usually the cursor is a black, non-scaled pointer,
which is easy to get lost on a dark background.

DMZ White looks much more visible in these contexts. It turns
out when the OS-level cursor is configured separately, e.g.
in gnome tweak tool, it changes cursors managed by the native
WM _but not_ nix applications, with the strange effect of
seeing the tweak-tool-selected cursor in "native" apps, and
when the cursor is moved over a nix-managed app, it switches
to the nix default. This change makes nix apps have at least
a more visible cursor. After applying this change, running
nix apps must be reloaded/restarted for the effect to become
visible.